restore conda plumbing (#1687) - minus formatted files

+++ package/src/common/dependencies/dependencies.ts
@@ -67,26 +67,31 @@ export async function configureDependency(
   const archDep = dependency.architectureDependencies[Deno.build.arch];
   if (archDep) {
     const platformDep = archDep[Deno.build.os];
-    info(`Downloading ${dependency.name}`);
+    const vendor = Deno.env.get("QUARTO_VENDOR_BINARIES");
+    let targetFile = "";
+    if (vendor === undefined || vendor === "true") {
+      info(`Downloading ${dependency.name}`);
 
-    let targetFile;
-    try {
-      targetFile = await downloadBinaryDependency(
-        dependency,
-        platformDep,
-        config,
-      );
-    } catch (error) {
-      const msg =
-        `Failed to Download ${dependency.name}\nAre you sure that version ${dependency.version} of ${dependency.bucket} has been archived using './quarto-bld archive-bin-deps'?\n${error.message}`;
-      throw new Error(msg);
+      try {
+        targetFile = await downloadBinaryDependency(
+          dependency,
+          platformDep,
+          config,
+        );
+      } catch (error) {
+        const msg =
+          `Failed to Download ${dependency.name}\nAre you sure that version ${dependency.version} of ${dependency.bucket} has been archived using './quarto-bld archive-bin-deps'?\n${error.message}`;
+        throw new Error(msg);
+      }
     }
 
     info(`Configuring ${dependency.name}`);
     await platformDep.configure(targetFile);
 
-    info(`Cleaning up`);
-    Deno.removeSync(targetFile);
+    if (targetFile) {
+      info(`Cleaning up`);
+      Deno.removeSync(targetFile);
+    }
   } else {
     throw new Error(
       `The architecture ${Deno.build.arch} is missing the dependency ${dependency.name}`,
